requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
werkzeug
gunicorn
